einige Anpassungen nach Test in SET

+++ de.setsoftware.reviewtool.core/src/de/setsoftware/reviewtool/ui/views/ReviewContentView.java
@@ -4,6 +4,7 @@ import java.io.File;
 
 import org.eclipse.core.resources.IMarker;
 import org.eclipse.core.runtime.CoreException;
+import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.jface.viewers.ITreeContentProvider;
 import org.eclipse.jface.viewers.LabelProvider;
 import org.eclipse.jface.viewers.TreeViewer;
@@ -15,6 +16,7 @@ import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Event;
 import org.eclipse.swt.widgets.Label;
 import org.eclipse.swt.widgets.Listener;
+import org.eclipse.swt.widgets.Shell;
 import org.eclipse.swt.widgets.Tree;
 import org.eclipse.swt.widgets.TreeItem;
 import org.eclipse.ui.IWorkbenchPage;
@@ -86,6 +88,12 @@ public class ReviewContentView extends ViewPart implements ReviewModeListener {
         try {
             final IWorkbenchPage page = PlatformUI.getWorkbench().getActiveWorkbenchWindow().getActivePage();
             final IMarker marker = SlicesInReview.createMarkerFor(new RealMarkerFactory(), fragment);
+            if (marker == null) {
+                final Shell shell = PlatformUI.getWorkbench().getActiveWorkbenchWindow().getShell();
+                MessageDialog.openError(shell, "Öffnen nicht möglich",
+                        "Öffnen über die Eclipse-Projekte nicht möglich für " + fragment);
+                return;
+            }
             IDE.openEditor(page, marker);
             marker.delete();
         } catch (final CoreException e) {
